move generic code to agent_windows.go
